apis Colocacao

+++ src/main/java/cv/dge/dge_api_intermed_lab/application/emprego/dto/CoordenadorOrientadorResponse.java
@@ -7,7 +7,7 @@ public record CoordenadorOrientadorResponse(
         String tipo,
         String tipoDesc,
         String nome,
-        Integer pessoaId,
+        Long pessoaId,
         String cargo,
         String email,
         String telemovel,
